Add package.json and UMD bootstrap
Prepare for publication on NPM by adding a package.json file and
wrapping in a script loader adapter that should work with AMD, CommonJS
and plain old script tags.

> Dear god why?

@jamesrwhite asked in issue #3 and I figured "why not". I'm still not
sure that anyone should ever use this code anywhere to do anything, but
sometimes practical needs get in the way of common sense.