Add support for Serializable interface
Add basic support for deserializing objects that implement the
Serializable interface. PHP serializes these objects as two strings, the
class name and the serialized payload. It is up to the class to
implement a method to parse the serialized payload and restore the
object state. Since we are disconnected from the PHP class
implementation in this javascript library we will return a dummy object
with the extracted data:

    {
      "__PHP_Incomplete_Class_Name": <class name>,
      "serialized": <serialized data>
    }